+++ multitude/Poetic/Utils.cpp
@@ -153,12 +153,12 @@ namespace Poetic
     if(ws.empty())
       return;
 
-    int n = ws.size();
+    size_t n = ws.size();
 
     std::vector<float> advances;
     advances.resize(n, 0);
 
-    fnt.advanceList(ws.c_str(), & advances[0], n);
+    fnt.advanceList(ws.c_str(), & advances[0], static_cast<int> (n));
 
     int lineStart = 0;
     int okEnd = 0;
